app/frontend/requirements.txt
azure-core==1.32.0
azure-identity==1.19.0
azure-storage-blob==12.23.1
azure-monitor-opentelemetry==1.6.4
requests==2.32.3
streamlit==1.41.1
